fix runner, use dtslint
