--- test/types/runner.ts
@@ -1,6 +1,0 @@
-/// <reference path="../../compromise.d.ts" />
-// import * as nlp from '../../src'
-import * as nlp from "../../builds/compromise";
-
-const r = nlp('he is fun', {})
-r.debug()
